requirements.txt
streamlit>=1.28.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.7
fpdf>=1.7.2
